requirements.txt
Flask>=2.2.0
pandas>=2.2.0
numpy>=1.24.0
schedule
tqdm


# 核心依赖
tenacity>=8.2.0             # 重试机制（指数退避）
sqlalchemy>=2.0.0           # ORM数据库操作
python-dotenv>=1.0.0        # 环境变量配置管理
# exchange-calendars>=4.5.0   # 交易日历（A股/港股/美股，Issue #373）


# from daily_stock_analysis 仓库 学习多源策略 /data_provider 文件夹
# 数据源依赖（多源策略，按优先级排序）
efinance>=0.5.5             # Priority 0: 东方财富数据源（最高优先级）https://github.com/Micro-sheep/efinance
akshare>=1.12.0             # Priority 1: 东方财富爬虫数据源
tushare>=1.4.0              # Priority 2: 挖地兔 Pro API
pytdx>=1.72                 # Priority 2: 通达信行情服务器
baostock>=0.8.0             # Priority 3: 证券宝数据
yfinance>=0.2.0             # Priority 4: Yahoo Finance (Fallback)



# 网络请求
requests>=2.31.0            # HTTP 请求
fake-useragent>=1.4.0       # 随机 User-Agent 防封禁